Add uname to client, various fixes
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@411 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/sha1.h
@@ -2,15 +2,15 @@
  *  sha1.h
  *
  *  Description:
- *	This is the header file for code which implements the Secure
- *	Hashing Algorithm 1 as defined in FIPS PUB 180-1 published
- *	April 17, 1995.
+ *      This is the header file for code which implements the Secure
+ *      Hashing Algorithm 1 as defined in FIPS PUB 180-1 published
+ *      April 17, 1995.
  *
- *	Many of the variable names in this code, especially the
- *	single character names, were used because those were the names
- *	used in the publication.
+ *      Many of the variable names in this code, especially the
+ *      single character names, were used because those were the names
+ *      used in the publication.
  *
- *	Please read the file sha1.c for more information.
+ *      Please read the file sha1.c for more information.
  *
  * Full Copyright Statement
  *
@@ -55,10 +55,10 @@
 /*
  * If you do not have the ISO standard stdint.h header file, then you
  * must typdef the following:
- *    name		meaning
- *  uint32_t	     unsigned 32 bit integer
- *  uint8_t	     unsigned 8 bit integer (i.e., unsigned char)
- *  int_least16_t    integer of >= 16 bits
+ *    name              meaning
+ *  uint32_t         unsigned 32 bit integer
+ *  uint8_t          unsigned 8 bit integer (i.e., unsigned char)
+ *  int32_t          integer of 32 bits
  *
  */
 
@@ -67,9 +67,9 @@
 enum
 {
     shaSuccess = 0,
-    shaNull,		/* Null pointer parameter */
-    shaInputTooLong,	/* input data too long */
-    shaStateError	/* called Input after Result */
+    shaNull,            /* Null pointer parameter */
+    shaInputTooLong,    /* input data too long */
+    shaStateError       /* called Input after Result */
 };
 #endif
 #define SHA1HashSize 20
@@ -82,15 +82,15 @@ typedef struct SHA1Context
 {
     uint32_t Intermediate_Hash[SHA1HashSize/4]; /* Message Digest  */
 
-    uint32_t Length_Low;	    /* Message length in bits	   */
-    uint32_t Length_High;	    /* Message length in bits	   */
+    uint32_t Length_Low;            /* Message length in bits      */
+    uint32_t Length_High;           /* Message length in bits      */
 
-			       /* Index into message block array   */
-    int_least16_t Message_Block_Index;
-    uint8_t Message_Block[64];	    /* 512-bit message blocks	   */
+                               /* Index into message block array   */
+    int32_t Message_Block_Index;
+    uint8_t Message_Block[64];      /* 512-bit message blocks      */
 
-    int Computed;		/* Is the digest computed?	   */
-    int Corrupted;	       /* Is the message digest corrupted? */
+    int Computed;               /* Is the digest computed?         */
+    int Corrupted;             /* Is the message digest corrupted? */
 } SHA1Context;
 
 /*
@@ -99,9 +99,9 @@ typedef struct SHA1Context
 
 int SHA1Init(SHA1Context *);
 int SHA1Update(SHA1Context *,
-	       const uint8_t *,
-	       unsigned int);
+               const uint8_t *,
+               unsigned int);
 int SHA1Final(SHA1Context *,
-	       uint8_t Message_Digest[SHA1HashSize]);
+               uint8_t Message_Digest[SHA1HashSize]);
 
 #endif
